refactor(compose): split catalog metadata
