+++ npm/compose/core/src/catalog-utilities.ts
@@ -1,0 +1,273 @@
+import { allTargets } from "./catalog-entries.ts";
+import type { ComposableUtilityMetadata } from "./catalog-schema.ts";
+
+const reactiveTargets = allTargets;
+const browserTargets = ["web", "desktop"] as const;
+
+/** Operational metadata for every public runtime utility. */
+export const utilities = [
+  {
+    name: "anyAbortSignal",
+    entry: "./abort-signal",
+    category: "async",
+    stability: "experimental",
+    targets: allTargets,
+    ssr: "safe",
+    hydration: "not-applicable",
+    cleanupOwners: ["returned-value"],
+    runtimeGlobals: ["AbortController", "AbortSignal"],
+    dependencies: [],
+  },
+  {
+    name: "timeoutAbortSignal",
+    entry: "./abort-signal",
+    category: "timing",
+    stability: "experimental",
+    targets: allTargets,
+    ssr: "safe",
+    hydration: "not-applicable",
+    cleanupOwners: ["returned-value"],
+    runtimeGlobals: ["AbortController", "AbortSignal", "DOMException", "globalThis"],
+    dependencies: ["anyAbortSignal"],
+  },
+  {
+    name: "deadlineAbortSignal",
+    entry: "./abort-signal",
+    category: "timing",
+    stability: "experimental",
+    targets: allTargets,
+    ssr: "safe",
+    hydration: "not-applicable",
+    cleanupOwners: ["returned-value"],
+    runtimeGlobals: [],
+    dependencies: ["timeoutAbortSignal"],
+  },
+  {
+    name: "useAsyncResource",
+    entry: "./async-resource",
+    category: "async",
+    stability: "experimental",
+    targets: reactiveTargets,
+    ssr: "safe",
+    hydration: "stable",
+    cleanupOwners: ["caller", "reactive-scope"],
+    runtimeGlobals: ["AbortController", "DOMException"],
+    dependencies: ["tryOnScopeDispose"],
+  },
+  {
+    name: "availableCapability",
+    entry: "./capability",
+    category: "capability",
+    stability: "experimental",
+    targets: allTargets,
+    ssr: "safe",
+    hydration: "not-applicable",
+    cleanupOwners: ["none"],
+    runtimeGlobals: [],
+    dependencies: [],
+  },
+  {
+    name: "unavailableCapability",
+    entry: "./capability",
+    category: "capability",
+    stability: "experimental",
+    targets: allTargets,
+    ssr: "safe",
+    hydration: "not-applicable",
+    cleanupOwners: ["none"],
+    runtimeGlobals: [],
+    dependencies: [],
+  },
+  {
+    name: "isCapabilityAvailable",
+    entry: "./capability",
+    category: "capability",
+    stability: "experimental",
+    targets: allTargets,
+    ssr: "safe",
+    hydration: "not-applicable",
+    cleanupOwners: ["none"],
+    runtimeGlobals: [],
+    dependencies: [],
+  },
+  {
+    name: "isCapabilityUnavailable",
+    entry: "./capability",
+    category: "capability",
+    stability: "experimental",
+    targets: allTargets,
+    ssr: "safe",
+    hydration: "not-applicable",
+    cleanupOwners: ["none"],
+    runtimeGlobals: [],
+    dependencies: [],
+  },
+  {
+    name: "createDisposalScope",
+    entry: "./disposal-scope",
+    category: "lifecycle",
+    stability: "experimental",
+    targets: reactiveTargets,
+    ssr: "safe",
+    hydration: "not-applicable",
+    cleanupOwners: ["caller", "reactive-scope"],
+    runtimeGlobals: [],
+    dependencies: ["tryOnScopeDispose"],
+  },
+  {
+    name: "useEventListener",
+    entry: "./event-listener",
+    category: "events",
+    stability: "experimental",
+    targets: reactiveTargets,
+    ssr: "safe",
+    hydration: "stable",
+    cleanupOwners: ["caller", "reactive-scope"],
+    runtimeGlobals: [],
+    dependencies: ["tryOnScopeDispose"],
+  },
+  {
+    name: "useLocale",
+    entry: "./locale",
+    category: "i18n",
+    stability: "experimental",
+    targets: reactiveTargets,
+    ssr: "deterministic-fallback",
+    hydration: "caller-managed",
+    cleanupOwners: ["none"],
+    runtimeGlobals: ["Intl", "navigator"],
+    dependencies: [],
+  },
+  {
+    name: "useMediaQuery",
+    entry: "./media-query",
+    category: "media",
+    stability: "experimental",
+    targets: browserTargets,
+    ssr: "deterministic-fallback",
+    hydration: "caller-managed",
+    cleanupOwners: ["reactive-scope"],
+    runtimeGlobals: ["window"],
+    dependencies: [],
+  },
+  {
+    name: "useReducedMotion",
+    entry: "./media-query",
+    category: "media",
+    stability: "experimental",
+    targets: browserTargets,
+    ssr: "deterministic-fallback",
+    hydration: "caller-managed",
+    cleanupOwners: ["reactive-scope"],
+    runtimeGlobals: ["window"],
+    dependencies: ["useMediaQuery"],
+  },
+  {
+    name: "tryOnScopeDispose",
+    entry: "./scope",
+    category: "lifecycle",
+    stability: "experimental",
+    targets: reactiveTargets,
+    ssr: "safe",
+    hydration: "not-applicable",
+    cleanupOwners: ["reactive-scope"],
+    runtimeGlobals: [],
+    dependencies: [],
+  },
+  {
+    name: "useTemporalNow",
+    entry: "./temporal",
+    category: "timing",
+    stability: "experimental",
+    targets: reactiveTargets,
+    ssr: "deterministic-fallback",
+    hydration: "caller-managed",
+    cleanupOwners: ["reactive-scope"],
+    runtimeGlobals: ["globalThis", "window"],
+    dependencies: [],
+  },
+  {
+    name: "useTemporalZonedDateTime",
+    entry: "./temporal",
+    category: "timing",
+    stability: "experimental",
+    targets: reactiveTargets,
+    ssr: "deterministic-fallback",
+    hydration: "caller-managed",
+    cleanupOwners: ["reactive-scope"],
+    runtimeGlobals: ["globalThis", "window"],
+    dependencies: ["useTemporalNow"],
+  },
+  {
+    name: "useCounter",
+    entry: "./use-counter",
+    category: "state",
+    stability: "experimental",
+    targets: reactiveTargets,
+    ssr: "safe",
+    hydration: "stable",
+    cleanupOwners: ["none"],
+    runtimeGlobals: [],
+    dependencies: [],
+  },
+  {
+    name: "useDebounced",
+    entry: "./use-debounced",
+    category: "timing",
+    stability: "experimental",
+    targets: reactiveTargets,
+    ssr: "deterministic-fallback",
+    hydration: "stable",
+    cleanupOwners: ["caller", "reactive-scope"],
+    runtimeGlobals: ["globalThis", "window"],
+    dependencies: ["tryOnScopeDispose"],
+  },
+  {
+    name: "useHistory",
+    entry: "./use-history",
+    category: "state",
+    stability: "experimental",
+    targets: reactiveTargets,
+    ssr: "safe",
+    hydration: "stable",
+    cleanupOwners: ["reactive-scope"],
+    runtimeGlobals: [],
+    dependencies: ["tryOnScopeDispose"],
+  },
+  {
+    name: "usePrevious",
+    entry: "./use-previous",
+    category: "state",
+    stability: "experimental",
+    targets: reactiveTargets,
+    ssr: "safe",
+    hydration: "stable",
+    cleanupOwners: ["none"],
+    runtimeGlobals: [],
+    dependencies: [],
+  },
+  {
+    name: "useThrottled",
+    entry: "./use-throttled",
+    category: "timing",
+    stability: "experimental",
+    targets: reactiveTargets,
+    ssr: "deterministic-fallback",
+    hydration: "stable",
+    cleanupOwners: ["caller", "reactive-scope"],
+    runtimeGlobals: ["globalThis", "window"],
+    dependencies: ["tryOnScopeDispose"],
+  },
+  {
+    name: "useToggle",
+    entry: "./use-toggle",
+    category: "state",
+    stability: "experimental",
+    targets: reactiveTargets,
+    ssr: "safe",
+    hydration: "stable",
+    cleanupOwners: ["none"],
+    runtimeGlobals: [],
+    dependencies: [],
+  },
+] as const satisfies readonly ComposableUtilityMetadata[];
